YT Trns lng/requirements.txt
dataclasses
google.generativeai
streamlit
langgraph.graph
typing
youtube_transcript_api
urllib.parse
